The crawler now stores and reloads state in case the crawl has been interrupted. For example due to networking errors or ctrl+c
To do so, a state file is created in the current working directory prior to shutdown of the crawler. If this file exists while initializing the crawler, it will be loaded. In case there are more endpoints than mentioned in the state-file, the crawler will _NOT_ overwrite / delete these new information.

Furthermore the generator is extended to use CREATE TABLE IF NOT EXISTS instead of simple CREATE TABLE statements. Also instead of printing some information while running, a logger is used now.